apps/api/requirements.txt
# HydroQ-QC-Assistant API Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Validation & Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Geospatial
rasterio>=1.3.9
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
fiona>=1.9.0

# ML & Scientific
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
scikit-image>=0.22.0
pandas>=2.2.0

# S-102 / HDF5 Export
h5py>=3.10.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
structlog>=24.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Development
ruff>=0.2.0
mypy>=1.8.0
